データ型修正

+++ db/migrate/20231021230814_create_post_comments.rb
@@ -1,8 +1,8 @@
 class CreatePostComments < ActiveRecord::Migration[6.1]
   def change
     create_table :post_comments do |t|
-      t.references :customer, null: false, type: :integer, foreign_key: true
-      t.references :post, null: false, type: :integer, foreign_key: true
+      t.references :customer, null: false, type: :bigint, foreign_key: true
+      t.references :post, null: false, type: :bigint, foreign_key: true
       t.text :comment, null: false
 
       t.timestamps
