+++ db/migrate/20231027020221_create_favorites.rb
@@ -1,8 +1,8 @@
 class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
-      t.integer :customer_id
-      t.integer :post_id
+      t.bigint :customer_id
+      t.bigint :post_id
 
       t.timestamps
     end
